Moved find commands to Viewer
